+++ src/kakkoiichris/stackvm/util/Extensions.kt
@@ -2,15 +2,15 @@ package kakkoiichris.stackvm.util
 
 import kotlin.math.absoluteValue
 
-fun Float.truncate() =
-    if (this - toInt() == 0F)
+fun Double.truncate() =
+    if (this - toInt() == 0.0)
         toInt().toString()
     else
         toString()
 
-val Float.bool get() = this != 0F
+val Double.bool get() = this != 0.0
 
-val Boolean.float get() = if (this) 1F else 0F
+val Boolean.float get() = if (this) 1.0 else 0.0
 
 fun Int.toAddress() =
     if (this < 0)
